data-processor/requirements.txt
scikit-learn
pandas
numpy
joblib
psycopg2-binary
sentence-transformers
torch
flask
flask-cors
python-dotenv
